+++ backend-EasyPark/src/main/java/com/backend/EasyPark/repository/EnderecoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.backend.EasyPark.entities.Endereco;
 
 @Repository
-public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
+public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
     
     List<Endereco> findByCep(String cep);
 }
