+++ backend-EasyPark/src/main/java/com/backend/EasyPark/repository/FabricanteRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.backend.EasyPark.entities.Fabricante;
 
 @Repository
-public interface FabricanteRepository extends JpaRepository<Fabricante, Long> {
+public interface FabricanteRepository extends JpaRepository<Fabricante, Integer> {
     
     Fabricante findByMarca(String marca);
