+++ backend-EasyPark/src/main/java/com/backend/EasyPark/repository/VeiculoRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.backend.EasyPark.entities.Veiculo;
 
 @Repository
-public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
+public interface VeiculoRepository extends JpaRepository<Veiculo, Integer> {
 
     List<Veiculo> findByOcupandoVagaTrue();
     // Método adicional que pode ser útil
